Refactor image decoding and rendering modules
